+++ Sources/BFKit/Linux/BFKit/BFApp.swift
@@ -33,6 +33,7 @@ import Foundation
 
 #if os(iOS)
     /// Get AppDelegate. To use it, cast to AppDelegate with "as! AppDelegate".
+    @available(*, deprecated: 3.0, message: "`appDelegate` will be removed in BFKit-Swift 3.0.")
     public let appDelegate: UIApplicationDelegate? = UIApplication.shared.delegate
 #endif