+++ jazzy.sh
@@ -2,7 +2,7 @@
 
 # Creates documentation using Jazzy.
 
-FRAMEWORK_VERSION=2.5.0
+FRAMEWORK_VERSION=2.6.0
 
 jazzy \
   --clean \