+++ CHANGELOG.md
@@ -56,6 +56,7 @@ All notable changes to this project will be documented in this file.<br>
 - Fixed missing style in `init(frame:, style:, cellSeparatorStyle:, separatorInset:, dataSource:, delegate:)` in UITableView extension
 - Fixed `animate(path:, count: = default, duration:, autoreverses: = default)` autoreverses
 - Fixed BFButton with a default `fadeDuration` value
+- Fixed wrong default duration in `fade(label:, duration:, toAttributedString:)` in WKInterfaceController extension
 
 ---